requirements.txt
aiobotocore==2.15.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
astropy==6.1.4
astropy-iers-data==0.2024.10.21.0.33.21
attrs==24.2.0
av==13.1.0
awscli==1.35.11
boto3==1.35.45
botocore==1.35.36
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
docutils==0.16
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.2
multidict==6.1.0
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
psutil==6.1.0
pvl==1.3.2
pyasn1==0.6.1
pyerfa==2.0.1.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rsa==4.7.2
s3fs==2024.10.0
s3transfer==0.10.3
setuptools==75.2.0
six==1.16.0
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.16.0
